% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_id.R
\name{add_id}
\alias{add_id}
\title{Add id in each group}
\usage{
add_id(tbl, grp = "group", id = "id")
}
\arguments{
\item{tbl}{A dataframe}

\item{grp, id}{A string to specify the column of group and id}
}
\value{
A dataframe
}
\description{
Add id in each group
}
\examples{
brk <- "EOS"
tbl <- tibble::tibble(col=c(rep("a", 2), brk, rep("b", 3), brk, rep("c", 4), brk))
add_group(tbl, col = "col") |>
  add_id(id = "id_in_group")

}
