% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{add_series_no}
\alias{add_series_no}
\title{Add series no col according to match condition.}
\usage{
add_series_no(tbl, cond = "", new_col = "series_no", starts_with = 1)
}
\arguments{
\item{tbl}{A tibble or data.frame.}

\item{cond}{Condition to split series no.}

\item{new_col}{A string name of new column.}

\item{starts_with}{A integer.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A tibble, which include new_col as series no.
}\if{html}{\out{</div>}}
}
\description{
Add series no col according to match condition.
}
\examples{
\dontrun{
  tbl <- tibble::tibble(col=c(rep("a", 2), "sep", rep("b", 3), "sep", rep("c", 4), "sep"))
  cond <- ".$col == 'sep'"   # Use ".$'colname'" to identify column
  add_series_no(tbl, cond = cond, new_col = "series_no")
}

}
