% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{add_text_id}
\alias{add_text_id}
\title{Add id column into result of morphological analysis}
\usage{
add_text_id(tbl, method, brk = "BP")
}
\arguments{
\item{tbl}{A tibble or data.frame.}

\item{method}{A text. Method to use: "mecab", "ginza",
"sudachi_a", "sudachi_b", "sudachi_c", or "chamame".
"a", "b" and "c" specify the mode of splitting.
"a" split shortest, "b" middle and "c" longest.
See https://github.com/WorksApplications/Sudachi for detail.
"chamame" use https://chamame.ninjal.ac.jp/ and rvest.}

\item{brk}{A string of break point}
}
\value{
A data.frame with column "text_id".
}
\description{
Internal function for moranajp_all().
Add \code{text_id} column when there is brk ("BP").
"BP": Break Point of moranajp
}
