% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_up
# @examples}
\alias{clean_up
# @examples}
\alias{clean_mecab_local}
\title{Clean up result of morphological analyzed data frame}
\usage{
clean_mecab_local(df, ...)
}
\arguments{
\item{df}{A dataframe including result of morphological analysis.}

\item{...}{Extra arguments to internal fuctions.}

\item{use_common_data}{A logical. TRUE: use data(stop_words).}

\item{stop_words}{A string vector.}

\item{synonym_df}{A datarame including synonym word pairs.
The first column: replace from, the second: replace to.}

\item{synonym_from, synonym_to}{A string vector. Length of synonym_from and synonym_to
should be the same.}
}
\value{
A dataframe.
}
\description{
Clean up result of morphological analyzed data frame
}
