% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{delete_parenthesis}
\alias{delete_parenthesis}
\title{Delete parenthesis and its internals}
\usage{
delete_parenthesis(df, pos = "pos_1")
}
\arguments{
\item{df}{A dataframe analysed by MeCab}

\item{pos}{A string to specify (pos) position of sentence}
}
\value{
A dataframe
}
\description{
Delete parenthesis and its internals
}
\examples{
library(tidyverse)
data(review_mecab)
cols <- c("text_id", "\\\\u8868\\\\u5c64\\\\u5f62", "\\\\u54c1\\\\u8a5e", 
          "\\\\u54c1\\\\u8a5e\\\\u7d30\\\\u5206\\\\u985e1", "\\\\u539f\\\\u5f62") \%>\%
         unescape_utf()
review_sudachi_a \%>\%
  unescape_utf()
  dplyr::mutate(`:=`(text_id, as.numeric(text_id))) \%>\%
  dplyr::filter(text_id < 5) \%>\%
  dplyr::select(dplyr::all_of(cols)) \%>\%
  print(n=120) \%>\%
  delete_parenthesis(pos = unescape_utf("\u54c1\u8a5e\u7d30\u5206\u985e1") ) \%>\%
  print(n=120)

}
