% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigram.R
\name{draw_bigram_network}
\alias{draw_bigram_network}
\alias{bigram}
\alias{trigram}
\alias{bigram_depend}
\alias{bigram_network}
\alias{word_freq}
\alias{bigram_network_plot}
\title{Draw bigram network using morphological analysis data.}
\usage{
draw_bigram_network(df, draw = TRUE, ...)

bigram(df, group = "sentence", depend = FALSE, term_depend = NULL, ...)

trigram(df, group = "sentence")

bigram_depend(df, group = "sentence")

bigram_network(bigram, rand_seed = 12, threshold = 100, ...)

word_freq(df, big_net, ...)

bigram_network_plot(
  big_net,
  freq,
  ...,
  arrow_size = 5,
  circle_size = 5,
  text_size = 5,
  font_family = "",
  arrow_col = "darkgreen",
  circle_col = "skyblue",
  x_limits = NULL,
  y_limits = NULL,
  no_scale = FALSE
)
}
\arguments{
\item{df}{A dataframe including result of morphological analysis.}

\item{draw}{A logical.}

\item{...}{Extra arguments to internal functions.}

\item{group}{A string to specify sentence.}

\item{depend}{A logical.}

\item{term_depend}{A string of dependent terms column to use bigram.}

\item{bigram}{A result of bigram().}

\item{rand_seed}{A numeric.}

\item{threshold}{A numeric used as threshold for frequency of bigram.}

\item{big_net}{A result of bigram_network().}

\item{freq}{A numeric of word frequency in bigram_network.
Can be got using word_freq().}

\item{arrow_size, circle_size, text_size, }{A numeric.}

\item{font_family}{A string.}

\item{arrow_col, circle_col}{A string to specify arrow and circle color
in bigram network.}

\item{x_limits, y_limits}{A Pair of numeric to specify range.}

\item{no_scale}{A logical. FALSE: Not draw x and y axis.}
}
\value{
A list including df (input), bigram, freq (frequency) and
gg (ggplot2 object of bigram network plot).
}
\description{
Draw bigram network using morphological analysis data.
}
\examples{

sentences <- 50
len <- 30
n <- sentences * len
x <- letters
prob <- (length(x):1) ^ 3
df <- 
  tibble::tibble(
    lemma = sample(x = x, size = n, replace = TRUE, prob = prob),
    sentence = rep(seq(sentences), each = len))
draw_bigram_network(df)

}
