% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_iconv.R
\name{escape_japanese}
\alias{escape_japanese}
\title{Generate code like "stringi::stri_unescape_unicode(...)"}
\usage{
escape_japanese(x)
}
\arguments{
\item{x}{A string or vector of Japanese}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A string or vector 
}\if{html}{\out{</div>}}
}
\description{
Generate code like "stringi::stri_unescape_unicode(...)"
}
\examples{
stringi::stri_unescape_unicode("\\\\u8868\\\\u5c64\\\\u5f62") |>
  print() |>
  escape_japanese()

}
