% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_iconv.R
\name{iconv_x}
\alias{iconv_x}
\title{iconv x}
\usage{
iconv_x(x, iconv = "", reverse = FALSE)
}
\arguments{
\item{x}{A string vector or a tibble.}

\item{iconv}{A text. Convert encoding of MeCab output.
Default (""): don't convert.
"CP932_UTF-8": iconv(output, from = "Shift-JIS" to = "UTF-8")
"EUC_UTF-8"  : iconv(output, from = "eucjp", to = "UTF-8")
iconv is also used to convert input text before running MeCab.
"CP932_UTF-8": iconv(input, from =  "UTF-8", to = "Shift-JIS")}

\item{reverse}{A logical.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A string vector.
}\if{html}{\out{</div>}}
}
\description{
iconv x
}
