% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{mecab}
\alias{mecab}
\title{Morphological analysis for text}
\usage{
mecab(tbl, bin_dir, tmp_dir, fileEncoding)
}
\arguments{
\item{tbl}{Tibble or data.frame.}

\item{bin_dir}{Text，Directory of mecab.}

\item{tmp_dir}{Text，Temporary directory for text.}

\item{fileEncoding}{Text, Fileencoding in mecab. "EUC", "CP932" (shift_jis) or "UTF-8".}
}
\value{
Tibble.       Output of MeCab.
}
\description{
Using MeCab in morphological analysis
}
\examples{
# not run
# data(neko)
# bin_dir <- "c:/mecab/bin"  # input your environment
# fileEncoding <- "CP932"    # input your environment
# mecab(neko, bin_dir=bin_dir, fileEncoding=fileEncoding)

}
\seealso{
mecab_all()
}
