% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neko}
\alias{neko}
\title{The first part of 'I Am a Cat' by Soseki Natsume}
\format{
A data frame with 9 rows and 1 variable:
\describe{
\item{text}{Body text. Escaped by stringi::stri_escape_unicode().}
}
}
\usage{
neko
}
\description{
The first part of 'I Am a Cat' by Soseki Natsume
}
\examples{
data(neko)
neko |>
  unescape_utf()
}
\keyword{datasets}
