% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neko_sudachi_a}
\alias{neko_sudachi_a}
\alias{neko_sudachi_b}
\alias{neko_sudachi_c}
\title{Analyzed data of neko by Sudachi}
\format{
A data frame with 3130 rows and 9 variable:
\describe{
\item{text_id}{id}
\item{\\u8868\\u5c64\\u5f62}{result of Sudachi}
\item{\\u54c1\\u8a5e}{result of Sudachi}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e1}{result of Sudachi}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e2}{result of Sudachi}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e3}{result of Sudachi}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e4}{result of Sudachi}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e5}{result of Sudachi}
\item{\\u539f\\u5f62}{result of Sudachi}
}

A data frame with 3088 rows and 9 variable:

A data frame with 3080 rows and 9 variable:
}
\usage{
neko_sudachi_a

neko_sudachi_b

neko_sudachi_c
}
\description{
Sudachi: https://github.com/WorksApplications/Sudachi
}
\examples{
data(neko_sudachi_a)
neko_sudachi_a |>
  unescape_utf()
}
\keyword{datasets}
