% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{review}
\alias{review}
\title{Full text of review article}
\format{
A data frame with 457 rows and 4 variables:
\describe{
\item{text}{Body text. Escaped by stringi::stri_escape_unicode().
Body text. Escaped by stringi::stri_escape_unicode().
Citation is as below.
Matsumura et al. 2014.
Conditions and conservation for biodiversity of the semi-natural
grassland vegetation on rice paddy levees.
Vegetation Science, 31, 193-218.
doi = 10.15031/vegsci.31.193
https://www.jstage.jst.go.jp/article/vegsci/31/2/31_193/_article/-char/en
}
\item{chap}{chapter}
\item{sect}{section}
\item{para}{paragraph}
}
}
\usage{
review
}
\description{
Full text of review article
}
\examples{
data(review)
review |>
  unescape_utf()
}
\keyword{datasets}
