% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{review_chamame}
\alias{review_chamame}
\title{Analyzed data of review by chamame}
\format{
A data frame with  21125 rows and 10 variable
(column names are escaped by stringi::stri_escape_unicode(),
stringi::stri_unescape_unicode() or unescape_utf() will show Japanese)
\describe{
\item{text_id}{id}
\item{chap}{chapter}
\item{sect}{section}
\item{para}{paragraph}
\item{\\u8868\\u5c64\\u5f62}{result of chamame}
\item{\\u54c1\\u8a5e}{result of chamame}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e1}{result of chamame}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e2}{result of chamame}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e3}{result of chamame}
\item{\\u539f\\u5f62}{result of chamame}
}
}
\usage{
review_chamame
}
\description{
chamame: https://chamame.ninjal.ac.jp/index.html
}
\examples{
data(review_chamame)
review_chamame |>
  unescape_utf()
}
\keyword{datasets}
