% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{review_ginza}
\alias{review_ginza}
\title{Analyzed data of review by GiNZA}
\format{
A data frame with 19514 rows and 16 variable:
\describe{
\item{text_id}{id}
\item{chap}{chapter}
\item{sect}{section}
\item{para}{paragraph}
\item{id}{result of GiNZA}
\item{\\u8868\\u5c64\\u5f62}{result of GiNZA}
\item{\\u539f\\u5f62}{result of GiNZA}
\item{UD\\u54c1\\u8a5e\\u30bf\\u30b0}{result of GiNZA}
\item{\\u54c1\\u8a5e}{result of GiNZA}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e1}{result of GiNZA}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e2}{result of GiNZA}
\item{\\u5c5e\\u6027}{result of GiNZA}
\item{\\u4fc2\\u53d7\\u5143}{result of GiNZA}
\item{\\u4fc2\\u53d7\\u30bf\\u30b0}{result of GiNZA}
\item{\\u4fc2\\u53d7\\u30da\\u30a2}{result of GiNZA}
\item{\\u305d\\u306e\\u4ed6}{result of GiNZA}
}
}
\usage{
review_ginza
}
\description{
GiNZA: https://megagonlabs.github.io/ginza/
}
\examples{
data(review_ginza)
review_ginza |>
  unescape_utf()
}
\keyword{datasets}
