% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{review_sudachi_a}
\alias{review_sudachi_a}
\alias{review_sudachi_b}
\alias{review_sudachi_c}
\title{Analyzed data of review by Sudachi}
\format{
A data frame with 20100 rows and 12 variable:
\describe{
\item{text_id}{id}
\item{chap}{chapter}
\item{sect}{section}
\item{para}{paragraph}
\item{\\u8868\\u5c64\\u5f62}{result of Sudachi}
\item{\\u54c1\\u8a5e}{result of Sudachi}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e1}{result of Sudachi}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e2}{result of Sudachi}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e3}{result of Sudachi}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e4}{result of Sudachi}
\item{\\u54c1\\u8a5e\\u7d30\\u5206\\u985e5}{result of Sudachi}
\item{\\u539f\\u5f62}{result of Sudachi}
}

A data frame with 19565 rows and 12 variable:

A data frame with 19526 rows and 12 variable:
}
\usage{
review_sudachi_a

review_sudachi_b

review_sudachi_c
}
\description{
Sudachi: https://github.com/WorksApplications/Sudachi
}
\examples{
data(review_sudachi_a)
review_sudachi_a |>
  unescape_utf()
}
\keyword{datasets}
