% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synonym}
\alias{synonym}
\title{An example of synonym word pairs}
\format{
A data frame with 25 rows and 2 variables:
\describe{
\item{from}{
Words to be replaced from.
Escaped by stringi::stri_escape_unicode().
}
\item{to}{
Words to be replaced to.
}
}
}
\usage{
synonym
}
\description{
An example of synonym word pairs
}
\examples{
data(synonym)
synonym |>
  unescape_utf()
}
\keyword{datasets}
