% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_id.R
\name{text_id_with_break}
\alias{text_id_with_break}
\alias{add_text_id_df}
\title{Add ids.}
\usage{
text_id_with_break(x, brk, end_with_brk = TRUE)

add_text_id_df(df, col, brk, end_with_brk = TRUE)
}
\arguments{
\item{x}{A string vector.}

\item{brk}{A string to specify the break between ids.}

\item{end_with_brk}{A logical.
TRUE: brk means the end of groups.
FALSE: brk means the beginning of groups.}

\item{df}{A dataframe.}

\item{col}{A string to specify the column.}
}
\value{
id_with_break() returns id vector, add_id_df() returns dataframe.
}
\description{
Add ids.
}
\examples{
tmp <- c("a", "brk", "b", "brk", "c")
brk <- "brk"
text_id_with_break(tmp, brk)
add_text_id_df(tibble::tibble(tmp), col = "tmp", "brk")

}
