% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_iconv.R
\name{unescape_utf}
\alias{unescape_utf}
\alias{escape_utf}
\title{Wrapper functions for escape and unescape unicode}
\usage{
unescape_utf(x)

escape_utf(x)
}
\arguments{
\item{x}{A dataframe or character vector}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A dataframe or character vector
}\if{html}{\out{</div>}}
}
\description{
Wrapper functions for escape and unescape unicode
}
\examples{
data(review_mecab)
review_mecab |>
  print() |>
  unescape_utf() |>
  print() |>
  escape_utf()

}
