% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_id.R
\name{add_sentence_no}
\alias{add_sentence_no}
\title{Wrapper function for add_group() to add sentence id}
\usage{
add_sentence_no(df, s_id = "sentence")
}
\arguments{
\item{df}{A dataframe}

\item{s_id}{A string for sentence colame}
}
\value{
A dataframe
}
\description{
Wrapper function for add_group() to add sentence id
}
\examples{
review_mecab |>
  unescape_utf() |>
  add_sentence_no() |>
  print(n=200)

}
