% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_id.R
\name{add_sentence_no}
\alias{add_sentence_no}
\title{Add sentence id}
\usage{
add_sentence_no(df, snt = "sentence")
}
\arguments{
\item{df}{A dataframe}

\item{snt}{A string for sentence colame}
}
\value{
A dataframe
}
\description{
Add sentence id
}
\examples{
review_mecab \%>\%
  unescape_utf() \%>\%
  add_sentence_no()

}
