% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{add_series_no}
\alias{add_series_no}
\title{Add series no col according to match condition.}
\usage{
add_series_no(tbl, cond = "", new_col = "series_no")
}
\arguments{
\item{tbl}{A tibble or data.frame.}

\item{cond}{Condition to split series no.}

\item{new_col}{A string name of new column.}
}
\value{
\preformatted{   A tibble, which include new_col as series no. 
}
}
\description{
Add series no col according to match condition.
}
