% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{add_text_id}
\alias{add_text_id}
\title{Add id column into result of morphological analysis}
\usage{
add_text_id(tbl, text_id = "text_id")
}
\arguments{
\item{tbl}{Tibble or data.frame.}
}
\value{
Tibble.
}
\description{
internal function for mecab_all()
}
