% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{add_text_id}
\alias{add_text_id}
\title{Add id column into result of morphological analysis}
\usage{
add_text_id(tbl, id = "text_id")
}
\arguments{
\item{tbl}{A tibble or data.frame.}

\item{id}{A text. Colnames for id of text.}
}
\value{
A tibble.
}
\description{
Internal function for mecab_all().
"EOF" means breaks of the end of a result in morphological analysis.
}
