% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{add_text_id}
\alias{add_text_id}
\title{Add id column into result of morphological analysis}
\usage{
add_text_id(tbl)
}
\arguments{
\item{tbl}{A tibble or data.frame.}
}
\description{
Internal function for moranajp_all().
In mecab: add_text_id() add \code{text_id} column when there is 'EOS'.
'EOS' means breaks of text in this package (and most of morphological analysis).
In ginza: add_text_id() add \code{text_id} column when starts with '#'
}
