% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{add_text_id}
\alias{add_text_id}
\title{Add id column into result of morphological analysis}
\usage{
add_text_id(tbl)
}
\arguments{
\item{tbl}{A tibble or data.frame.}
}
\value{
\preformatted{   A tibble.
}
}
\description{
Internal function for mecab_all().
'EOS' means breaks of the end of a result in morphological analysis.
add_text_id() add 1 to \code{text_id} column when there is 'EOS' & NA.
}
