% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up.R
\name{combine_words}
\alias{combine_words}
\alias{combi_words}
\title{Combine words after morphological analysis}
\usage{
combine_words(df, text = "text_col", combi, sep = "-")

combi_words(x, combi, sep = "-")
}
\arguments{
\item{df}{A dataframe including result of morphological analysis.}

\item{text}{A text. Colnames for morphological analysis.}

\item{combi}{A string (combi_words()) or string vector (combine_words()) to combine words.}

\item{sep}{A string of separator of words}
}
\description{
Combine words after morphological analysis
}
\examples{
library(tidyverse)
x <- letters[1:10]
combi <- c("b-c")
combi_words(x, combi)
expected <- c("a", "bc", NA, "d", "e", "f", "g", "h", "i", "j")
testthat::expect_equal(combi_words(x, combi), expected)

df <- unescape_utf(review_chamame) \%>\% head(20)
combi <- c("生物-多様", "農地-は", "農産-物", "生産-する")
combine_words(df, "原形", combi)

}
