% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_group.R
\name{make_groups}
\alias{make_groups}
\alias{make_groups_sub}
\alias{max_sum_str_length}
\title{Make groups by splitting string length}
\usage{
make_groups(
  tbl,
  text_col = "text",
  length = 8000,
  tmp_group = "tmp_group",
  str_length = "str_length"
)

make_groups_sub(tbl, text_col, n_group, tmp_group, str_length)

max_sum_str_length(tbl, tmp_group, str_length)
}
\arguments{
\item{tbl}{A tibble or data.frame.}

\item{text_col}{A text. Colnames for morphological analysis.}

\item{length}{A numeric.}

\item{tmp_group, str_length}{A string to use temporary.}

\item{n_group}{A numeric.}
}
\value{
A tibble.   Output of morphological analysis and added column "text_id".
\code{NULL} when \code{method = "chamame"} and
web chamame is not available.

A string

A string

A string

A character vector

A character vector

A character vector

A character vector

A character vector

A data.frame
}
\description{
Using 'MeCab' for morphological analysis.
Keep other colnames in dataframe.
}
\examples{
# make_groups() is an internal function, called by moranajp_all().
data(review)
review |>
  unescape_utf() |>
  moranajp:::make_groups(length = 1000)
}
