% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_make_group.R
\name{make_groups}
\alias{make_groups}
\alias{make_groups_sub}
\alias{max_str_length}
\title{Make groups by splitting string length}
\usage{
make_groups(tbl, text_col = "text", length = 8000)

make_groups_sub(tbl, n_group)

max_str_length(tbl)
}
\arguments{
\item{n_group}{A numeric.}
}
\description{
Make groups by splitting string length
}
