% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{mecab_all}
\alias{mecab_all}
\alias{mecab}
\alias{make_mecab_cmd}
\alias{out_cols_mecab}
\title{Morphological analysis for a speciefic column in dataframe}
\usage{
mecab_all(tbl, text_col = "text", bin_dir = "", fileEncoding = "CP932")

mecab(tbl, bin_dir, fileEncoding)

make_mecab_cmd(bin_dir, input, output)

out_cols_mecab()
}
\arguments{
\item{tbl}{A tibble or data.frame.}

\item{text_col}{A text. Colnames for morphological analysis.}

\item{bin_dir}{A text. Directory of mecab.}

\item{fileEncoding}{A text. Fileencoding in mecab.
'EUC', 'CP932' (shift_jis) or 'UTF-8'.}

\item{input, output}{A text. File path of input and output.}
}
\value{
Tibble. Output of MeCab.
}
\description{
Using MeCab in morphological analysis
Keep other colnames in dataframe
}
\examples{
# not run
# library(tidyverse)
# bin_dir <- 'd:/pf/mecab/bin/'  # input your environment
# fileEncoding <- 'CP932'        # input your environment
# data(neko)
# neko <- 
#   neko \%>\%
#   dplyr::mutate(text=stringi::stri_unescape_unicode(text)) \%>\%
#   dplyr::mutate(cols=1:nrow(.))
# mecab_all(neko, 
#           text_col='text', 
#           bin_dir=bin_dir, 
#           fileEncoding=fileEncoding) \%>\%
#   print(n=100)
}
\seealso{
mecab()
}
