% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{mecab_all}
\alias{mecab_all}
\alias{mecab}
\title{Morphological analysis for a speciefic column in dataframe}
\usage{
mecab_all(
  tbl,
  text_col = "text",
  bin_dir = ".",
  tmp_dir = NULL,
  fileEncoding = "CP932"
)

mecab(tbl, bin_dir, tmp_dir, fileEncoding)
}
\arguments{
\item{tbl}{Tibble or data.frame.}

\item{text_col}{Text. Colnames for morphological analysis.}

\item{bin_dir}{Text，Directory of mecab.}

\item{tmp_dir}{Text，Temporary directory for text.}

\item{fileEncoding}{Text, Fileencoding in mecab.
"EUC", "CP932" (shift_jis) or "UTF-8".}
}
\value{
Tibble. Output of MeCab.
}
\description{
Using MeCab in morphological analysis
Keep other colnames in dataframe
}
\examples{
# not run
# data(neko)
# bin_dir <- "c:/mecab/bin/" # input your environment
# fileEncoding <- "CP932"    # input your environment
# neko \%>\%
#   tibble::tibble(text=., cols=rep(1:2, each=2))
# mecab_all(neko, 
#           text_col="text", 
#           bin_dir=bin_dir, 
#           fileEncoding=fileEncoding) \%>\%
#   print(n=nrow(.))
}
\seealso{
mecab()
}
