% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neko}
\alias{neko}
\title{The first part of 'I Am a Cat' by Soseki Natsume}
\format{
A tibble
\describe{
\item{text}{Body text.}
}
}
\usage{
neko
}
\description{
The first part of 'I Am a Cat' by Soseki Natsume
}
\keyword{datasets}
