% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neko_ginza}
\alias{neko_ginza}
\title{Analyzed data of neko by GiNZA}
\format{
A data frame with 733 rows and 14 variable:
\describe{
\item{text_id    }{result of GiNZA}
\item{id         }{result of GiNZA}
\item{form       }{result of GiNZA}
\item{lemma      }{result of GiNZA}
\item{upos       }{result of GiNZA}
\item{xpos       }{result of GiNZA}
\item{pos_1      }{result of GiNZA}
\item{pos_2      }{result of GiNZA}
\item{pos_3      }{result of GiNZA}
\item{feats      }{result of GiNZA}
\item{head       }{result of GiNZA}
\item{deprel     }{result of GiNZA}
\item{deps       }{result of GiNZA}
\item{misc       }{result of GiNZA}
}
}
\usage{
neko_ginza
}
\description{
GiNZA: https://megagonlabs.github.io/ginza/
}
\examples{
data(neko_ginza)
neko_ginza \%>\%
  unescape_utf()
}
\keyword{datasets}
