% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranajp.R
\name{remove_brk}
\alias{remove_brk}
\title{Remove break point and other unused rows from the result of
morphological analysis}
\usage{
remove_brk(tbl, method, brk = "BPMJP")
}
\arguments{
\item{tbl}{A tibble or data.frame.}

\item{method}{A text. Method to use: "mecab", "ginza",
"sudachi_a", "sudachi_b", "sudachi_c", or "chamame".
"a", "b" and "c" specify the mode of splitting.
"a" split shortest, "b" middle and "c" longest.
See https://github.com/WorksApplications/Sudachi for detail.
"chamame" use https://chamame.ninjal.ac.jp/ and rvest.}

\item{brk}{A string of break point}
}
\value{
A data.frame.
}
\description{
Internal function for moranajp_all().
}
